recommendations/requirements.txt
APScheduler==3.10.4
joblib==1.3.2
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
scikit-surprise==1.1.3
scipy==1.12.0
six==1.16.0
SQLAlchemy==2.0.28
surprise==0.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
